# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация API Filmotek",
        "title": "Filmotek API",
        "contact": {
            "name": "Поддержка API",
            "url": "https://github.com/shuklarituparn",
            "email": "rtprnshukla@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/movies/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Фильмы"
                ],
                "summary": "Получить все фильмы с пагинацией",
                "operationId": "read-all-movies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество элементов на странице",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Поле для сортировки (по умолчанию рейтинг)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Порядок сортировки (ASC или DESC, по умолчанию DESC)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фильмов",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReadAllMoviesResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный размер страницы или номер страницы",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/movies/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Фильмы"
                ],
                "summary": "Создать новый фильм",
                "operationId": "create-movie",
                "parameters": [
                    {
                        "description": "Объект фильма для создания",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMovieModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Фильм успешно создан",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateMovieResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/movies/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Фильмы"
                ],
                "summary": "Удалить фильм",
                "operationId": "delete-movie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор фильма",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм успешно удален",
                        "schema": {
                            "$ref": "#/definitions/controllers.DeleteMovieResponse"
                        }
                    },
                    "400": {
                        "description": "Требуется идентификатор фильма",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не удалось удалить фильм или его связи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/movies/get": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Фильмы"
                ],
                "summary": "Получить фильм по идентификатору",
                "operationId": "read-movie-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор фильма",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Подробности фильма",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReadMovieResponse"
                        }
                    },
                    "400": {
                        "description": "Требуется идентификатор фильма",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Фильм не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении фильма",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/movies/patch": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Фильмы"
                ],
                "summary": "Частично обновить существующий фильм",
                "operationId": "patch-movie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор фильма",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные для частичного обновления фильма",
                        "name": "patchData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMovieModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/controllers.PatchMovieResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не удалось обновить фильм или его связи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/movies/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Фильмы"
                ],
                "summary": "Обновить существующий фильм",
                "operationId": "update-movie",
                "parameters": [
                    {
                        "description": "Объект фильма для обновления",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMovieModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateMovieResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не удалось обновить фильм",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/login": {
            "post": {
                "description": "Вход пользователя с использованием электронной почты и пароля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginUserModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход в систему",
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверная электронная почта или пароль",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/refresh": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление доступа и обновления токенов JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Обновление токенов JWT",
                "responses": {
                    "200": {
                        "description": "Новые токены доступа и обновления",
                        "schema": {
                            "$ref": "#/definitions/controllers.RefreshTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный или истекший токен",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/register": {
            "post": {
                "description": "Зарегистрировать нового пользователя с помощью электронной почты и пароля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Зарегистрировать нового пользователя",
                "parameters": [
                    {
                        "description": "Информация о пользователе",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterUserModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateMovieResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Movie"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "controllers.DeleteMovieResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "controllers.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "controllers.PatchMovieResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controllers.ReadAllMoviesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Movie"
                    }
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "controllers.ReadMovieResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Movie"
                }
            }
        },
        "controllers.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "refresh": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UpdateMovieResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Movie"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Actor": {
            "type": "object",
            "required": [
                "birth_date",
                "first_name",
                "gender",
                "last_name"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Movie"
                    }
                }
            }
        },
...
        "models.CreateMovieModel": {
            "type": "object",
            "required": [
                "rating",
                "title"
            ],
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "genre": {
                    "type": "string"
                },
                "rating": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0
                },
                "release_year": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
        "models.LoginUserModel": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "models.Movie": {
            "type": "object",
            "required": [
                "genre",
                "rating",
                "release_year",
                "title"
            ],
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0
                },
                "release_year": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
        "models.RegisterUserModel": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "models.UpdateMovieModel": {
            "type": "object",
            "required": [
                "id",
                "rating",
                "title"
            ],
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0
                },
                "release_year": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}